# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Duration:
    description: |-
      A Duration represents the elapsed time between two instants
      as an int64 nanosecond count. The representation limits the
      largest representable duration to approximately 290 years.
    format: int64
    type: integer
    x-go-package: time
  ExpiringTask:
    properties:
      Complete:
        type: boolean
      CreatedAt:
        format: date-time
        type: string
      DeletedAt:
        format: date-time
        type: string
      Description:
        type: string
      ExpiringAt:
        format: date-time
        type: string
      ID:
        format: uint64
        type: integer
      PersonID:
        format: int64
        type: integer
      ProjectID:
        format: int64
        type: integer
      TimeLeft:
        $ref: '#/definitions/Duration'
      Title:
        type: string
      UpdatedAt:
        format: date-time
        type: string
    type: object
    x-go-package: github.com/psebaraj/gogetitdone/models
...
  PriorityLevelType:
    format: uint64
    type: integer
    x-go-package: github.com/psebaraj/gogetitdone/models
  PriorityTask:
    properties:
      Complete:
        type: boolean
      CreatedAt:
        format: date-time
        type: string
      DeletedAt:
        format: date-time
        type: string
      Description:
        type: string
      ID:
        format: uint64
        type: integer
      PersonID:
        format: int64
        type: integer
      PriorityLevel:
        $ref: '#/definitions/PriorityLevelType'
      ProjectID:
        format: int64
        type: integer
      Title:
        type: string
      UpdatedAt:
        format: date-time
        type: string
    type: object
    x-go-package: github.com/psebaraj/gogetitdone/models
  Project:
    properties:
      CreatedAt:
        format: date-time
        type: string
      DeletedAt:
        format: date-time
        type: string
      ExpiringTasks:
        items:
          $ref: '#/definitions/ExpiringTask'
        type: array
      ID:
        format: uint64
        type: integer
      PersonID:
        format: int64
        type: integer
      PriorityTasks:
        items:
          $ref: '#/definitions/PriorityTask'
        type: array
      Tasks:
        items:
          $ref: '#/definitions/Task'
        type: array
      Title:
        type: string
      UpdatedAt:
        format: date-time
        type: string
    type: object
    x-go-package: github.com/psebaraj/gogetitdone/models
  Task:
    properties:
      Complete:
        type: boolean
      CreatedAt:
        format: date-time
        type: string
      DeletedAt:
        format: date-time
        type: string
      Description:
        type: string
      ID:
        format: uint64
        type: integer
      PersonID:
        format: int64
        type: integer
      ProjectID:
        format: int64
        type: integer
      Title:
        type: string
      UpdatedAt:
        format: date-time
        type: string
    type: object
    x-go-package: github.com/psebaraj/gogetitdone/models
info:
  description: Documentation of the GoGetItDone endpoints of the backend
  title: GoGetItDone
  version: 1.0.0
paths:
...
  /changedescrition/prioritytask/{id}:
    patch:
      consumes:
      - application/json
      operationId: changePriorityTaskDescription
      produces:
      - application/json
      responses:
        "200":
          description: PriorityTask
          schema:
            $ref: '#/definitions/PriorityTask'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - PriorityTask
  /changedescrition/task/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeTaskDescription
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /changeproject/prioritytask/{id}:
    patch:
      consumes:
      - application/json
      operationId: changePriorityTaskProject
      produces:
      - application/json
      responses:
        "200":
          description: PriorityTask
          schema:
            $ref: '#/definitions/PriorityTask'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - PriorityTask
  /changeproject/task/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeTaskProject
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /changetitle/prioritytask/{id}:
    patch:
      consumes:
      - application/json
      operationId: changePriorityTaskTitle
      produces:
      - application/json
      responses:
        "200":
          description: PriorityTask
          schema:
            $ref: '#/definitions/PriorityTask'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - PriorityTask
  /changetitle/task/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeTaskTitle
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /complete/prioritytask/{id}:
    patch:
      operationId: togglePriorityTaskComplete
      produces:
      - application/json
      responses:
        "200":
          description: PriorityTask
          schema:
            $ref: '#/definitions/PriorityTask'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - PriorityTask
  /complete/task/{id}:
    patch:
      operationId: toggleTaskComplete
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /create/prioritytask:
    post:
      consumes:
      - application/json
      operationId: createPriorityTask
      produces:
      - application/json
      responses:
        "201":
          description: PriorityTask
          schema:
            $ref: '#/definitions/PriorityTask'
        "400":
          $ref: '#/responses/nil'
        "507":
          $ref: '#/responses/nil'
      tags:
      - PriorityTask
  /create/project:
    post:
      consumes:
      - application/json
      operationId: createProject
      produces:
      - application/json
      responses:
        "201":
          description: Project
          schema:
            $ref: '#/definitions/Project'
        "507":
          $ref: '#/responses/nil'
      tags:
      - Project
...
  /delete/prioritytask/{id}:
    delete:
      operationId: deletePriorityTask
      produces:
      - application/json
      responses:
        "200":
          description: PriorityTask
          schema:
            $ref: '#/definitions/PriorityTask'
        "404":
          $ref: '#/responses/nil'
        "500":
          $ref: '#/responses/nil'
      tags:
      - PriorityTask
  /delete/project/{id}:
    delete:
      operationId: deleteProject
      produces:
      - application/json
      responses:
        "200":
          description: Project
          schema:
            $ref: '#/definitions/Project'
        "404":
          $ref: '#/responses/nil'
        "500":
          $ref: '#/responses/nil'
      tags:
      - Project
...
  /prioritytask/{id}:
    get:
      operationId: getPriorityTask
      produces:
      - application/json
      responses:
        "200":
          description: PriorityTask
          schema:
            $ref: '#/definitions/PriorityTask'
        "404":
          $ref: '#/responses/nil'
      tags:
      - PriorityTask
  /prioritytask/plevel/{id}/{new_level}:
    patch:
      operationId: changePriorityLevel
      produces:
      - application/json
      responses:
        "200":
          description: PriorityTask
          schema:
            $ref: '#/definitions/PriorityTask'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - PriorityTask
  /prioritytasks:
    get:
      operationId: getPriorityTasks
      produces:
      - application/json
      responses:
        "200":
          description: PriorityTask
          schema:
            items:
              $ref: '#/definitions/PriorityTask'
            type: array
        "204":
          $ref: '#/responses/nil'
      tags:
      - PriorityTask
  /project/{id}:
    get:
      operationId: getProject
      produces:
      - application/json
      responses:
        "200":
          description: Project
          schema:
            $ref: '#/definitions/Project'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Project
  /projects/{email}:
    get:
      operationId: getProjects
      produces:
      - application/json
      responses:
        "200":
          description: Project
          schema:
            items:
              $ref: '#/definitions/Project'
            type: array
        "204":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Project
...
produces:
- application/json
...
schemes:
- http
swagger: "2.0"